Split out a reusable runInlineScripts utility function and handle attributes correctly
